common: substitute "off_t" with "long long int"
Modify libpmemlog and libpmemblk API to not use "off_t" type for
function arguments or returned value.

This modification is dictated by incompatible "off_t" definitions on
64-bit Linux and Windows operating systems.  On Windows, "off_t" is
always 32-bit long, which is not enough to hold offsets of large (>4GB)
pmem pool files.
To avoid potential problems with redefining "off_t" or with adding
additional typedefs, seems like the best choice is to use "long long"
type.

+++ src/examples/libpmemblk/manpage.c
@@ -43,7 +43,7 @@
 #include <libpmemblk.h>
 
 /* size of the pmemblk pool -- 1 GB */
-#define	POOL_SIZE ((off_t)(1 << 30))
+#define	POOL_SIZE ((size_t)(1 << 30))
 
 /* size of each element in the pmem pool */
 #define	ELEMENT_SIZE 1024
